Added a frame for the scheduler class, and another class
I added a new class for testing the simulation using the format
provided on the assignment. One thing i excluded was the time variable. I will
incorporate it after but for now i just wanted to make sure i could
establish a connection with the scheduler class. Also please send
packets to port 1111 and receive packets from port 2222. Don't use the
local ip address for sending because when we run the assignment, we will
have the three systems running on different machines not locally.